Removed calls for ffmpeg for OpenCASCADE.  Removed nlohmann_json (now included with current).  Removed all references to nlohmann.  Updating douninst.sh for md2roff to remove link when package is uninstalled.

+++ Development/md2roff/douninst.sh
@@ -5,4 +5,7 @@
 # NOTE: This script is run AFTER package removal, so be careful!
 #       Consider it optional, use if it is really needed.
 #
-rm /usr/bin/roff2pdf
+
+if [ -f /usr/bin/roff2pdf ]; then
+	rm /usr/bin/roff2pdf
+fi
